+++ firmware/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Update Rust to 1.98.0.
 - Support both STM32G474 and STM32G473 targets.
 - Remove VPD loading built-ins.
+- Sleep between interrupt using WFI.
 
 ## v0.3.1